# add #[inline] to small and single-use functions