On ARM:
| bench name | current | new | diff |
|-|-|-|-|
| bench_check_raw_byte_str_ascii | 44621.22 ns/iter (+/- 67.52) | 32252.26 ns/iter (+/- 85.11) | -27.7% |
| bench_check_raw_c_str_ascii | 41249.71 ns/iter (+/- 459.54) | 30423.24 ns/iter (+/- 33.53) | -26.2% |
| bench_check_raw_c_str_non_ascii | 52466.45 ns/iter (+/- 78.41) | 52351.31 ns/iter (+/- 56.71) | -0.2% |
| bench_check_raw_c_str_unicode | 174486.0 ns/iter (+/- 1116.62) | 166608.7 ns/iter (+/- 1539.8) | -4.5% |
| bench_check_raw_str_ascii | 40627.54 ns/iter (+/- 87.79) | 28335.81 ns/iter (+/- 41.45) | -30.3% |
| bench_check_raw_str_non_ascii | 50431.16 ns/iter (+/- 82.04) | 50323.16 ns/iter (+/- 72.71) | -0.2% |
| bench_check_raw_str_unicode | 168052.34 ns/iter (+/- 728.17) | 157957.16 ns/iter (+/- 602.8) | -6.0% |
| bench_skip_ascii_whitespace | 13112.82 ns/iter (+/- 56.07) | 12063.44 ns/iter (+/- 52.15) | -8.0% |
| bench_unescape_byte_str_ascii | 69061.89 ns/iter (+/- 201.68) | 69282.8 ns/iter (+/- 215.7) | 0.3% |
| bench_unescape_byte_str_ascii_escape | 81271.92 ns/iter (+/- 224.8) | 81490.03 ns/iter (+/- 182.96) | 0.3% |
| bench_unescape_byte_str_hex_escape | 125719.84 ns/iter (+/- 295.52) | 124372.93 ns/iter (+/- 421.28) | -1.1% |
| bench_unescape_byte_str_mixed_escape | 348058.69 ns/iter (+/- 3154.79) | 346675.38 ns/iter (+/- 2412.19) | -0.4% |
| bench_unescape_c_str_ascii | 81765.96 ns/iter (+/- 158.23) | 78098.67 ns/iter (+/- 152.0) | -4.5% |
| bench_unescape_c_str_ascii_escape | 96378.31 ns/iter (+/- 1333.76) | 99110.02 ns/iter (+/- 151.25) | 2.8% |
| bench_unescape_c_str_hex_escape_ascii | 161821.62 ns/iter (+/- 672.23) | 151202.48 ns/iter (+/- 301.01) | -6.6% |
| bench_unescape_c_str_hex_escape_byte | 162515.88 ns/iter (+/- 375.19) | 145952.07 ns/iter (+/- 348.79) | -10.2% |
| bench_unescape_c_str_mixed_escape | 1046189.8 ns/iter (+/- 3429.65) | 1022389.5 ns/iter (+/- 2655.44) | -2.3% |
| bench_unescape_c_str_non_ascii | 110622.88 ns/iter (+/- 286.26) | 96923.48 ns/iter (+/- 351.63) | -12.4% |
| bench_unescape_c_str_unicode | 385263.4 ns/iter (+/- 1329.9) | 347683.62 ns/iter (+/- 1072.2) | -9.8% |
| bench_unescape_c_str_unicode_escape | 308760.33 ns/iter (+/- 1787.27) | 312357.17 ns/iter (+/- 646.29) | 1.2% |
| bench_unescape_str_ascii | 77105.57 ns/iter (+/- 163.34) | 71903.65 ns/iter (+/- 183.93) | -6.7% |
| bench_unescape_str_ascii_escape | 105557.74 ns/iter (+/- 219.26) | 102428.44 ns/iter (+/- 748.51) | -3.0% |
| bench_unescape_str_hex_escape | 165474.25 ns/iter (+/- 457.63) | 165896.69 ns/iter (+/- 557.95) | 0.3% |
| bench_unescape_str_mixed_escape | 906218.8 ns/iter (+/- 2277.11) | 904109.3 ns/iter (+/- 4160.06) | -0.2% |
| bench_unescape_str_non_ascii | 98056.04 ns/iter (+/- 178.18) | 98091.26 ns/iter (+/- 182.63) | 0.0% |
| bench_unescape_str_unicode | 343870.1 ns/iter (+/- 1217.51) | 343175.53 ns/iter (+/- 1244.8) | -0.2% |
| bench_unescape_str_unicode_escape | 596445.35 ns/iter (+/- 2718.8) | 594864.6 ns/iter (+/- 9608.78) | -0.3% |

Someone please check x86_64!

BTW I made some small improvements (I hope) to the bench script (mostly in `get_good_results`):

<details>

```python
import subprocess


class Bench:
    def __init__(self, line):
        self.name = line.split(' ')[1]
        self.speed = float(line.split('... bench:')[1].split('/iter')[0].strip().split(' ')[0].replace(',', ''))
        self.margin = float(line.split('(+/-')[1].split(')')[0].strip().replace(',', ''))


def run_bench():
    res = subprocess.run(["cargo", "+nightly", "bench"], check=True, stdout=subprocess.PIPE)
    return res.stdout.decode('utf-8')


def get_benches():
    print("> Retrieving benches...")
    out = run_bench()
    print("> Done")

    lines = out.split("\n")
    i = 0
    found_bench = False
    while i < len(lines):
        line = lines[i].strip()
        i += 1
        if line.startswith("test result:"):
            found_bench = True
            break
    if not found_bench:
        raise "Not found benchmarks in:\n{}".format(out)
    benches = {}
    while i < len(lines):
        line = lines[i].strip()
        i += 1
        if line.startswith("test result:"):
            break
        if not line.startswith("test bench_"):
            continue
        bench = Bench(line)
        benches[bench.name] = bench
    return benches


def get_good_results():
    benches = get_benches()

    for i in range(10):
        new_benches = get_benches()
        worst_rel_margin_1000 = 0
        updates = 0
        for key in benches.keys():
            if new_benches[key].speed < benches[key].speed and new_benches[key].margin < 2 * benches[key].margin:
                benches[key] = new_benches[key]
                updates += 1
            rel_margin_1000 = 1000 * benches[key].margin / benches[key].speed
            if rel_margin_1000 > worst_rel_margin_1000:
                worst_rel_margin_1000 = rel_margin_1000
        if worst_rel_margin_1000 > 5:
            print(">> Worst relative margin (iteration {}, updates: {}): {}".format(
                i, updates, worst_rel_margin_1000))
        else:
            break
    return benches


def get_str_for_bench(bench):
    return "{} ns/iter (+/- {})".format(bench.speed, bench.margin)


def show_results(benches):
    print("==== BENCHES RESULTS =====")
    for value in benches.values():
        print("{}: {}".format(value.name, get_str_for_bench(value)))


def show_comparison(current, new_ones):
    print("=== COMPARISON ===")
    print("| bench name | current | new | diff |")
    print("|-|-|-|-|")
    for key in current.keys():
        print("| {} | {} | {} | {:.1f}% |".format(
            key,
            get_str_for_bench(current[key]),
            get_str_for_bench(new_ones[key]),
            (new_ones[key].speed * 100. / current[key].speed) - 100.,
        ))


def main():
    subprocess.run(["git", "checkout", "main"], check=True)
#    subprocess.run(["git", "checkout", "inline-main"], check=True)
#    subprocess.run(["git", "checkout", "nonzero-for-cstr"], check=True)
    current = get_good_results()
    show_results(current)
    
#    subprocess.run(["git", "checkout", "inline"], check=True)
    subprocess.run(["git", "checkout", "inline-main"], check=True)
#    subprocess.run(["git", "checkout", "nonzero-for-cstr"], check=True)
    new_ones = get_good_results()
    show_results(new_ones)
    
    show_comparison(current, new_ones)


if __name__ == "__main__":
    main()
```

</details>