@GuillaumeGomez 2025-06-24
Thanks!

Here are the results:

| bench name | current | new | diff |
|-|-|-|-|
| bench_check_raw_byte_str_ascii | 21091.37 ns/iter (+/- 306.46) | 23836.23 ns/iter (+/- 148.15) | 13.0% |
| bench_check_raw_c_str_ascii | 20228.55 ns/iter (+/- 308.63) | 24671.73 ns/iter (+/- 76.81) | 22.0% |
| bench_check_raw_c_str_non_ascii | 37315.17 ns/iter (+/- 354.43) | 38698.32 ns/iter (+/- 93.38) | 3.7% |
| bench_check_raw_c_str_unicode | 115163.2 ns/iter (+/- 196.06) | 123726.65 ns/iter (+/- 1754.97) | 7.4% |
| bench_check_raw_str_ascii | 19828.49 ns/iter (+/- 216.05) | 24592.26 ns/iter (+/- 280.61) | 24.0% |
| bench_check_raw_str_non_ascii | 35885.5 ns/iter (+/- 338.35) | 37088.98 ns/iter (+/- 104.94) | 3.4% |
| bench_check_raw_str_unicode | 109651.19 ns/iter (+/- 455.53) | 119296.16 ns/iter (+/- 1830.07) | 8.8% |
| bench_skip_ascii_whitespace | 6308.74 ns/iter (+/- 21.53) | 6315.13 ns/iter (+/- 19.25) | 0.1% |
| bench_unescape_byte_str_ascii | 43806.82 ns/iter (+/- 499.42) | 43580.15 ns/iter (+/- 122.83) | -0.5% |
| bench_unescape_byte_str_ascii_escape | 49234.48 ns/iter (+/- 379.34) | 49047.87 ns/iter (+/- 139.34) | -0.4% |
| bench_unescape_byte_str_hex_escape | 78803.46 ns/iter (+/- 870.99) | 77621.41 ns/iter (+/- 177.82) | -1.5% |
| bench_unescape_byte_str_mixed_escape | 217997.52 ns/iter (+/- 2631.37) | 217488.67 ns/iter (+/- 702.22) | -0.2% |
| bench_unescape_c_str_ascii | 54373.06 ns/iter (+/- 232.85) | 44254.3 ns/iter (+/- 170.22) | -18.6% |
| bench_unescape_c_str_ascii_escape | 66438.86 ns/iter (+/- 98.75) | 59793.99 ns/iter (+/- 529.13) | -10.0% |
| bench_unescape_c_str_hex_escape_ascii | 105934.43 ns/iter (+/- 821.54) | 85070.95 ns/iter (+/- 178.62) | -19.7% |
| bench_unescape_c_str_hex_escape_byte | 106402.87 ns/iter (+/- 714.71) | 85044.34 ns/iter (+/- 167.69) | -20.1% |
| bench_unescape_c_str_mixed_escape | 650502.8 ns/iter (+/- 9717.05) | 557822.4 ns/iter (+/- 7040.85) | -14.2% |
| bench_unescape_c_str_non_ascii | 76833.2 ns/iter (+/- 925.05) | 62701.57 ns/iter (+/- 169.42) | -18.4% |
| bench_unescape_c_str_unicode | 262831.82 ns/iter (+/- 972.51) | 213803.75 ns/iter (+/- 625.26) | -18.7% |
| bench_unescape_c_str_unicode_escape | 185955.28 ns/iter (+/- 1732.51) | 163486.75 ns/iter (+/- 641.41) | -12.1% |
| bench_unescape_str_ascii | 45088.81 ns/iter (+/- 324.68) | 45127.8 ns/iter (+/- 84.89) | 0.1% |
| bench_unescape_str_ascii_escape | 67136.64 ns/iter (+/- 437.04) | 65999.89 ns/iter (+/- 418.29) | -1.7% |
| bench_unescape_str_hex_escape | 103124.07 ns/iter (+/- 296.07) | 101489.16 ns/iter (+/- 927.54) | -1.6% |
| bench_unescape_str_mixed_escape | 742458.35 ns/iter (+/- 4591.73) | 729269.3 ns/iter (+/- 5127.74) | -1.8% |
| bench_unescape_str_non_ascii | 63100.61 ns/iter (+/- 156.42) | 63314.37 ns/iter (+/- 526.48) | 0.3% |
| bench_unescape_str_unicode | 217618.15 ns/iter (+/- 1489.77) | 216865.7 ns/iter (+/- 1687.46) | -0.3% |
| bench_unescape_str_unicode_escape | 348800.22 ns/iter (+/- 1227.2) | 348860.6 ns/iter (+/- 1171.6) | 0.0% |