@hkBst 2025-06-24
`#[cfg_attr(any(target_arch = "aarch64", target_arch = "arm"), inline)]` to the rescue? Can you recheck x86-64?