@GuillaumeGomez 2025-06-24
Summary:

| bench name | x86_64 | ARM |
|-|-|-|
| bench_check_raw_byte_str_ascii | 2.5% | 0% |
| bench_check_raw_c_str_ascii | -0.7% | -0.4% |
| bench_check_raw_c_str_non_ascii | 0.0% | 0.2% |
| bench_check_raw_c_str_unicode | -0.0% | -0.5% |
| bench_check_raw_str_ascii | -0.2% | 0.1% |
| bench_check_raw_str_non_ascii | -0.1% | 0% |
| bench_check_raw_str_unicode | -0.1% | 0.1% |
| bench_skip_ascii_whitespace | 0.1% | 0.1% |
| bench_unescape_byte_str_ascii | -0.2% | 0.4% |
| bench_unescape_byte_str_ascii_escape | 0.2% | 0% |
| bench_unescape_byte_str_hex_escape | -1.3% | 0% |
| bench_unescape_byte_str_mixed_escape | -16.2% | -0.2% |
| bench_unescape_c_str_ascii | -30.0% | -6.2% |
| bench_unescape_c_str_ascii_escape | -9.9% | 1.5% |
| bench_unescape_c_str_hex_escape_ascii | -19.7% | -5% |
| bench_unescape_c_str_hex_escape_byte | -20.1% | -7.9% |
| bench_unescape_c_str_mixed_escape | -14.2% | -2.8% |
| bench_unescape_c_str_non_ascii | -18.2% | -13.7% |
| bench_unescape_c_str_unicode | -18.5% | -11.4% |
| bench_unescape_c_str_unicode_escape | -11.7% | 0.6% |
| bench_unescape_str_ascii | 0.2% | -5.1% |
| bench_unescape_str_ascii_escape | -1.0% | -2.3% |
| bench_unescape_str_hex_escape | -1.4% | -0.1% |
| bench_unescape_str_mixed_escape | -1.5% | -0.1% |
| bench_unescape_str_non_ascii | 0.5% | -0.1% |
| bench_unescape_str_unicode | 0.2% | -0.4% |
| bench_unescape_str_unicode_escape | 0.1% | 0.1% |

Seems all good to me, thanks for the performance improvement!